+++ test/unit/pulled_search/run_program.py
@@ -153,6 +153,7 @@ class LoggerTest(object):
 
         pass
 
+
 class CfgTest(object):
 
     """Class:  CfgTest