+++ test/unit/pulled_search/unit_test_run.sh
@@ -19,6 +19,7 @@ test/unit/pulled_search/non_processed.py
 test/unit/pulled_search/process_docid.py
 test/unit/pulled_search/process_files.py
 test/unit/pulled_search/process_insert.py
+test/unit/pulled_search/process_json.py
 test/unit/pulled_search/process_list.py
 test/unit/pulled_search/run_program.py
 test/unit/pulled_search/validate_dirs.py